www: Extract generic wizard view from apply RefuseWizardView
